BW-240: Complete Add Author functionality added. Moved ReservationDTO to right folder.
